Create remove-div-and-span-tags-from-article
Description: We who write articles using chat gpt or sometimes copy some topics from different websites and then rewrite the article. At that moment, some html tags were copied in the article without our knowledge. Actually these tags are of no use to us. But having this unwanted tag or code causes problem in seo of the article. Many times it also happens that the article is indexed in Google and even if it is indexed it does not get a good rank. If you post 100 articles and see that your article has a lot of unwanted code that you don't really need. But now how to remove this unnecessary code from so many articles. Thinking about this? After you install this plugin, click on the update button of the post from which you want to remove this unprovision code. See no more codes. If you activate this plugin, no matter how much necessary code there is inside the code, no matter what you do, publish, draft, pending, there will be no unnecessary code. And you think the post format (design) of your article would be right. This plugin will only remove <div class ="example"></div > and <span class ="example"> </span> tags. Keep the rest of the tags intact. For which the post format (design) of the article will be correct.